split into common and action-specific utils
