+++ src/actions/main/utils/constants.ts
@@ -14,5 +14,3 @@ export const EXPECTED_PERMISSIONS_BLOCK = [
 // the environment changes in a way that would cause a replay to behave differently, e.g. upgrading to a newer
 // replay-orchestrator-launcher version, or changing the version of puppeteer.
 export const LOGICAL_ENVIRONMENT_VERSION = 5;
-
-export const DOCS_URL = "https://app.meticulous.ai/docs/github-actions-v2";
