+++ src/common/check-connection.ts
