+++ src/common/constants.ts
@@ -1,0 +1,1 @@
+export const DOCS_URL = "https://app.meticulous.ai/docs/github-actions-v2";
