+++ src/common/environment.utils.ts
